+++ src/view/src/rocprofvis_charts.h
@@ -16,7 +16,7 @@ public:
     virtual float GetTrackHeight()                                                   = 0;
     virtual void  Render()                                                           = 0;
     virtual int   ReturnChartID()                                                    = 0;
-    virtual void  UpdateMovement(float zoom, float movement, float& min_x, float& max_x,
+    virtual void  UpdateMovement(float zoom, float movement, double& min_x, double& max_x,
                                  float scale_x, float m_scroll_position)             = 0;
     virtual const std::string& GetName()                                             = 0;
     virtual void  SetColorByValue(rocprofvis_color_by_value_t color_by_value_digits) = 0;
